LibWeb: Split GridTemplateAreaStyleValue out of StyleValue.{h,cpp}
